Extract AuditableEntity to its own file
